+++ ast/commandflag/flags.go
@@ -97,10 +97,11 @@ type GitCloneOpts struct {
 }
 
 type HealthCheckOpts struct {
-	Interval    time.Duration `long:"interval" description:"The interval between healthchecks" default:"30s"`
-	Timeout     time.Duration `long:"timeout" description:"The timeout before the command is considered failed" default:"30s"`
-	StartPeriod time.Duration `long:"start-period" description:"An initialization time period in which failures are not counted towards the maximum number of retries"`
-	Retries     int           `long:"retries" description:"The number of retries before a container is considered unhealthy" default:"3"`
+	Interval      time.Duration `long:"interval" description:"The interval between healthchecks" default:"30s"`
+	Timeout       time.Duration `long:"timeout" description:"The timeout before the command is considered failed" default:"30s"`
+	StartPeriod   time.Duration `long:"start-period" description:"An initialization time period in which failures are not counted towards the maximum number of retries"`
+	Retries       int           `long:"retries" description:"The number of retries before a container is considered unhealthy" default:"3"`
+	StartInterval time.Duration `long:"start-interval" description:"The time interval between health checks during the start period" default:"5s"`
 }
 
 type WithDockerOpts struct {