+++ buildcontext/gitlookup.go
@@ -40,6 +40,7 @@ type gitMatcher struct {
 	strictHostKeyChecking bool
 	port                  int
 	prefix                string
+	sshCommand            string
 }
 
 type gitProtocol string
@@ -124,7 +125,7 @@ func knownHostsToKeyScans(knownHosts string) []string {
 }
 
 // AddMatcher adds a new matcher for looking up git repos
-func (gl *GitLookup) AddMatcher(name, pattern, sub, user, password, prefix, suffix, protocol, knownHosts string, strictHostKeyChecking bool, port int) error {
+func (gl *GitLookup) AddMatcher(name, pattern, sub, user, password, prefix, suffix, protocol, knownHosts string, strictHostKeyChecking bool, port int, sshCommand string) error {
 	gl.mu.Lock()
 	defer gl.mu.Unlock()
 	p := gitProtocol(protocol)
@@ -158,6 +159,7 @@ func (gl *GitLookup) AddMatcher(name, pattern, sub, user, password, prefix, suff
 		suffix:                suffix,
 		protocol:              p,
 		strictHostKeyChecking: strictHostKeyChecking,
+		sshCommand:            sshCommand,
 	}
 
 	// update existing entry
@@ -481,19 +483,18 @@ func (gl *GitLookup) lookupNetRCCredential(host string) (login, password string,
 
 var errMakeCloneURLSubNotSupported = fmt.Errorf("makeCloneURL does not support gitMatcher substitution")
 
-func (gl *GitLookup) makeCloneURL(m *gitMatcher, host, gitPath string) (string, []string, error) {
+func (gl *GitLookup) makeCloneURL(m *gitMatcher, host, gitPath string) (gitURL string, keyScans []string, sshCommand string, err error) {
 	if m.sub != "" {
-		return "", nil, errMakeCloneURLSubNotSupported
+		return "", nil, "", errMakeCloneURLSubNotSupported
 	}
 
-	var err error
 	configuredProtocol := m.protocol
 	user := m.user
 	password := m.password
 	if configuredProtocol == autoProtocol {
 		configuredProtocol, err = gl.detectProtocol(host)
 		if err != nil {
-			return "", nil, err
+			return "", nil, "", err
 		}
 		switch configuredProtocol {
 		case sshProtocol:
@@ -504,8 +505,6 @@ func (gl *GitLookup) makeCloneURL(m *gitMatcher, host, gitPath string) (string, 
 		}
 	}
 
-	var gitURL string
-	var keyScans []string
 	switch configuredProtocol {
 	case sshProtocol:
 		if user == "" {
@@ -543,10 +542,10 @@ func (gl *GitLookup) makeCloneURL(m *gitMatcher, host, gitPath string) (string, 
 		}
 		_, keyScans, err = gl.getHostKeyAlgorithms(host)
 		if err != nil {
-			return "", nil, err
+			return "", nil, "", err
 		}
 		if len(keyScans) == 0 && m.strictHostKeyChecking {
-			return "", nil, errors.Errorf("no known_hosts entries exist for %s", host)
+			return "", nil, "", errors.Errorf("no known_hosts entries exist for %s", host)
 		}
 	case httpProtocol:
 		if user != "" || password != "" {
@@ -563,9 +562,9 @@ func (gl *GitLookup) makeCloneURL(m *gitMatcher, host, gitPath string) (string, 
 		}
 		gitURL = "https://" + userAndPass + host + "/" + gitPath
 	default:
-		return "", nil, errors.Errorf("unsupported protocol: %s", configuredProtocol)
+		return "", nil, "", errors.Errorf("unsupported protocol: %s", configuredProtocol)
 	}
-	return gitURL, keyScans, nil
+	return gitURL, keyScans, m.sshCommand, nil
 }
 
 // TODO eventually we should use gitutil.parseURL directly; but for now we want to avoid this change to keep this commit smaller
@@ -615,16 +614,18 @@ func parseGitProtocol(remote string) (string, int) {
 //	"github.com/earthly/earthly/examples/go" ---> ("git@github.com/earthly/earthly.git", "examples/go")
 //
 // Additionally a ssh keyscan might be returned (or an empty string indicating none was configured)
-func (gl *GitLookup) GetCloneURL(path string) (string, string, []string, error) {
+// Also, a custom "git ssh command" may be returned. This is part of this function since the user may
+// specify a command necessary to clone their repository successfully.
+func (gl *GitLookup) GetCloneURL(path string) (gitURL string, subPath string, keyScans []string, sshCommand string, err error) {
 	gl.mu.Lock()
 	defer gl.mu.Unlock()
 	match, m, err := gl.getGitMatcherByPath(path)
 	if err != nil {
-		return "", "", nil, err
+		return "", "", nil, "", err
 	}
 
 	n := len(match)
-	subPath := ""
+	subPath = ""
 	if len(path) > n {
 		subPath = path[n+1:]
 		path = path[:n]
@@ -632,9 +633,11 @@ func (gl *GitLookup) GetCloneURL(path string) (string, string, []string, error) 
 	host := path[:strings.IndexByte(path, '/')]
 	gitPath := m.prefix + match[(strings.IndexByte(match, '/')+1):] + m.suffix
 
+	sshCommand = m.sshCommand
+
 	if m.sub != "" {
 		if !m.re.MatchString(path) {
-			return "", "", nil, errors.Errorf("failed to determine git path to clone for %q", path)
+			return "", "", nil, "", errors.Errorf("failed to determine git path to clone for %q", path)
 		}
 		gitURL := m.re.ReplaceAllString(path, m.sub)
 		gl.console.VerbosePrintf("converted earthly reference %s to git url %s (using regex substitution %s)", path, stringutil.ScrubCredentials(gitURL), stringutil.ScrubCredentials(m.sub))
@@ -644,29 +647,28 @@ func (gl *GitLookup) GetCloneURL(path string) (string, string, []string, error) 
 			subHost := remote[:strings.IndexByte(remote, '/')]
 			_, keyScans, err = gl.getHostKeyAlgorithms(subHost)
 			if err != nil {
-				return "", "", nil, err
+				return "", "", nil, "", err
 			}
 			if len(keyScans) == 0 && m.strictHostKeyChecking {
-				return "", "", nil, errors.Errorf("no known_hosts entries exist for substituted host %s", subHost)
+				return "", "", nil, "", errors.Errorf("no known_hosts entries exist for substituted host %s", subHost)
 			}
 		}
-		return gitURL, subPath, keyScans, nil
+		return gitURL, subPath, keyScans, sshCommand, nil
 	}
 
-	gitURL, keyScans, err := gl.makeCloneURL(m, host, gitPath)
+	gitURL, keyScans, sshCommand, err = gl.makeCloneURL(m, host, gitPath)
 	if err != nil {
-		return "", "", nil, err
+		return "", "", nil, "", err
 	}
 	gl.console.VerbosePrintf("converted earthly reference %s to git url %s", path, stringutil.ScrubCredentials(gitURL))
-	return gitURL, subPath, keyScans, nil
+	return gitURL, subPath, keyScans, sshCommand, nil
 }
 
 // ConvertCloneURL takes a url such as https://github.com/user/repo.git or git@github.com:user/repo.git
 // and makes use of configured git credentials and protocol preferences to convert it into the appropriate
 // https or ssh protocol.
-// it also returns a keyScan
-func (gl *GitLookup) ConvertCloneURL(inURL string) (string, []string, error) {
-	var err error
+// it also returns a keyScan and sshCommand
+func (gl *GitLookup) ConvertCloneURL(inURL string) (gitURL string, keyScans []string, sshCommand string, err error) {
 	var host string
 	var gitPath string
 
@@ -675,7 +677,7 @@ func (gl *GitLookup) ConvertCloneURL(inURL string) (string, []string, error) {
 	case HTTPProtocol, HTTPSProtocol:
 		splits := strings.SplitN(remote, "/", 2)
 		if len(splits) != 2 {
-			return "", nil, errors.Errorf("failed to split path from host in %s", remote)
+			return "", nil, "", errors.Errorf("failed to split path from host in %s", remote)
 		}
 		host = splits[0]
 		gitPath = splits[1]
@@ -683,14 +685,14 @@ func (gl *GitLookup) ConvertCloneURL(inURL string) (string, []string, error) {
 		if sshutil.IsImplicitSSHTransport(inURL) {
 			splits := strings.SplitN(remote, ":", 2)
 			if len(splits) != 2 {
-				return "", nil, errors.Errorf("failed to split path from host in %s", remote)
+				return "", nil, "", errors.Errorf("failed to split path from host in %s", remote)
 			}
 			host = splits[0]
 			gitPath = splits[1]
 		} else {
 			u, err := url.Parse(inURL)
 			if err != nil {
-				return "", nil, errors.Wrapf(err, "failed to parse %s", inURL)
+				return "", nil, "", errors.Wrapf(err, "failed to parse %s", inURL)
 			}
 			if u.Scheme != "ssh" {
 				panic(fmt.Sprintf("expected scheme of ssh; got %s", u.Scheme)) // shouldn't happen
@@ -699,29 +701,28 @@ func (gl *GitLookup) ConvertCloneURL(inURL string) (string, []string, error) {
 			gitPath = u.Path
 		}
 	default:
-		return "", nil, errors.Errorf("unsupported git protocol %v", protocol)
+		return "", nil, "", errors.Errorf("unsupported git protocol %v", protocol)
 	}
 
 	m := gl.getGitMatcherByName(host)
 	if m.sub != "" {
 		path := host + strings.TrimSuffix(gitPath, ".git")
 		if !m.re.MatchString(path) {
-			return "", nil, errors.Errorf("failed to determine git path to clone for %q", path)
+			return "", nil, "", errors.Errorf("failed to determine git path to clone for %q", path)
 		}
-		gitURL := m.re.ReplaceAllString(path, m.sub)
-		var keyScans []string
+		gitURL = m.re.ReplaceAllString(path, m.sub)
 		remote, protocol := parseGitProtocol(gitURL)
 		if protocol == SSHProtocol {
 			subHost := remote[:strings.IndexByte(remote, '/')]
 			_, keyScans, err = gl.getHostKeyAlgorithms(subHost)
 			if err != nil {
-				return "", nil, err
+				return "", nil, "", err
 			}
 			if len(keyScans) == 0 && m.strictHostKeyChecking {
-				return "", nil, errors.Errorf("no known_hosts entries exist for substituted host %s", subHost)
+				return "", nil, "", errors.Errorf("no known_hosts entries exist for substituted host %s", subHost)
 			}
 		}
-		return gitURL, keyScans, nil
+		return gitURL, keyScans, m.sshCommand, nil
 	}
 
 	return gl.makeCloneURL(m, host,