+++ cmd/earthly/subcmd/build_cmd.go
@@ -25,6 +25,7 @@ import (
 	"github.com/moby/buildkit/session/socketforward/socketprovider"
 	"github.com/moby/buildkit/session/sshforward/sshprovider"
 	"github.com/moby/buildkit/util/entitlements"
+	buildkitgitutil "github.com/moby/buildkit/util/gitutil"
 	"github.com/pkg/errors"
 	"github.com/urfave/cli/v2"
 
@@ -62,7 +63,6 @@ import (
 	"github.com/earthly/earthly/util/syncutil/semutil"
 	"github.com/earthly/earthly/util/termutil"
 	"github.com/earthly/earthly/variables"
-	buildkitgitutil "github.com/moby/buildkit/util/gitutil"
 )
 
 const autoSkipPrefix = "auto-skip"
@@ -243,7 +243,8 @@ func (a *Build) ActionBuildImp(cliCtx *cli.Context, flagArgs, nonFlagArgs []stri
 		gitConfigEmail  string
 	)
 	if !target.IsRemote() {
-		if meta, err := gitutil.Metadata(cliCtx.Context, target.GetLocalPath(), a.cli.Flags().GitBranchOverride); err == nil {
+		meta, _ := gitutil.Metadata(cliCtx.Context, target.GetLocalPath(), a.cli.Flags().GitBranchOverride)
+		if meta != nil {
 			// Git commit detection here is best effort
 			gitCommitAuthor = meta.Author
 		}
@@ -251,7 +252,6 @@ func (a *Build) ActionBuildImp(cliCtx *cli.Context, flagArgs, nonFlagArgs []stri
 			gitConfigEmail = email
 		}
 	}
-
 	cleanCollection := cleanup.NewCollection()
 	defer cleanCollection.Close()
 
@@ -320,6 +320,7 @@ func (a *Build) ActionBuildImp(cliCtx *cli.Context, flagArgs, nonFlagArgs []stri
 
 	skipDB, targetHash, doSkip, err := a.initAutoSkip(cliCtx.Context, target, overridingVars, cloudClient)
 	if err != nil {
+		a.cli.Console().PrintFailure("auto-skip")
 		return err
 	}
 	if doSkip {
@@ -334,14 +335,14 @@ func (a *Build) ActionBuildImp(cliCtx *cli.Context, flagArgs, nonFlagArgs []stri
 		return errors.Wrapf(err, "could not init frontend")
 	}
 
-	// Collect info to help with printing a richer message in the beginning of the build or on failure to reserve satellite due to missing build minutes.
-	if err = a.cli.CollectBillingInfo(cliCtx.Context, cloudClient, a.cli.OrgName()); err != nil {
-		a.cli.Console().DebugPrintf("failed to get billing plan info, error is %v\n", err)
-	}
-
 	err = a.cli.ConfigureSatellite(cliCtx, cloudClient, gitCommitAuthor, gitConfigEmail)
 	if err != nil {
 		return errors.Wrapf(err, "could not configure satellite")
+	}
+
+	// Collect info to help with printing a richer message in the beginning of the build or on failure to reserve satellite due to missing build minutes.
+	if err = a.cli.CollectBillingInfo(cliCtx.Context, cloudClient, a.cli.OrgName()); err != nil {
+		a.cli.Console().DebugPrintf("failed to get billing plan info, error is %v\n", err)
 	}
 
 	// After configuring frontend and satellites, buildkit address should not be empty.
@@ -428,7 +429,7 @@ func (a *Build) ActionBuildImp(cliCtx *cli.Context, flagArgs, nonFlagArgs []stri
 		authChildren = append(authChildren, dockerauthprovider.NewDockerAuthProvider(cfg, nil).(auth.AuthServer))
 	}
 
-	authProvider := authprovider.New(authChildren)
+	authProvider := authprovider.New(a.cli.Console(), authChildren)
 	attachables = append(attachables, authProvider)
 
 	gitLookup := buildcontext.NewGitLookup(a.cli.Console(), a.cli.Flags().SSHAuthSock)
@@ -614,7 +615,7 @@ func (a *Build) ActionBuildImp(cliCtx *cli.Context, flagArgs, nonFlagArgs []stri
 		a.cli.Console().ColorPrintf(color.New(color.FgHiYellow), "Streaming logs to %s\n\n", logstreamURL)
 	}
 
-	a.maybePrintBuildMinutesInfo(a.cli.OrgName())
+	a.maybePrintBuildMinutesInfo(cliCtx)
 
 	_, err = b.BuildTarget(cliCtx.Context, target, buildOpts)
 	if err != nil {
@@ -669,7 +670,7 @@ func (a *Build) updateGitLookupConfig(gitLookup *buildcontext.GitLookup) error {
 		if suffix == "" {
 			suffix = ".git"
 		}
-		err := gitLookup.AddMatcher(k, pattern, v.Substitute, v.User, v.Password, v.Prefix, suffix, auth, v.ServerKey, common.IfNilBoolDefault(v.StrictHostKeyChecking, true), v.Port)
+		err := gitLookup.AddMatcher(k, pattern, v.Substitute, v.User, v.Password, v.Prefix, suffix, auth, v.ServerKey, common.IfNilBoolDefault(v.StrictHostKeyChecking, true), v.Port, v.SSHCommand)
 		if err != nil {
 			return errors.Wrap(err, "gitlookup")
 		}
@@ -831,47 +832,56 @@ func (a *Build) initAutoSkip(ctx context.Context, target domain.Target, overridi
 	console := a.cli.Console().WithPrefix(autoSkipPrefix)
 	consoleNoPrefix := a.cli.Console()
 
-	if a.cli.Flags().Push {
-		return nil, nil, false, errors.New("--push cannot be used with --auto-skip")
-	}
-
 	if a.cli.Flags().NoCache {
 		return nil, nil, false, errors.New("--no-cache cannot be used with --auto-skip")
 	}
 
-	var (
-		skipDB      bk.BuildkitSkipper
-		targetHash  []byte
-		orgName     string
-		projectName string
-	)
+	orgName := a.cli.Flags().OrgName
 
-	orgName, projectName, targetHash, err := inputgraph.HashTarget(ctx, inputgraph.HashOpt{
-		Target:           target,
-		Console:          a.cli.Console(),
-		CI:               a.cli.Flags().CI,
-		BuiltinArgs:      variables.DefaultArgs{EarthlyVersion: a.cli.Version(), EarthlyBuildSha: a.cli.GitSHA()},
-		OverridingVars:   overridingVars,
-		EarthlyCIRunner:  a.cli.Flags().EarthlyCIRunner,
-		SkipProjectCheck: a.cli.Flags().LocalSkipDB != "",
+	targetHash, err := inputgraph.HashTarget(ctx, inputgraph.HashOpt{
+		Target:          target,
+		Console:         a.cli.Console(),
+		CI:              a.cli.Flags().CI,
+		BuiltinArgs:     variables.DefaultArgs{EarthlyVersion: a.cli.Version(), EarthlyBuildSha: a.cli.GitSHA()},
+		OverridingVars:  overridingVars,
+		EarthlyCIRunner: a.cli.Flags().EarthlyCIRunner,
 	})
 	if err != nil {
-		return nil, nil, false, errors.Wrapf(err, "unable to calculate hash for %s", target)
+		return nil, nil, false, errors.Wrapf(err, "auto-skip is unable to calculate hash for %s", target)
 	}
 
-	skipDB, err = bk.NewBuildkitSkipper(a.cli.Flags().LocalSkipDB, orgName, projectName, target.GetName(), client)
+	if a.cli.Flags().LocalSkipDB == "" && orgName == "" {
+		orgName, _, err = inputgraph.ParseProjectCommand(ctx, target, console)
+		if err != nil {
+			return nil, nil, false, errors.New("organization not found in Earthfile, command flag or environmental variables")
+		}
+	}
+
+	if !target.IsRemote() {
+		meta, err := gitutil.Metadata(ctx, target.GetLocalPath(), a.cli.Flags().GitBranchOverride)
+		if err != nil {
+			console.VerboseWarnf("unable to detect all git metadata: %v", err.Error())
+		}
+		target = gitutil.ReferenceWithGitMeta(target, meta).(domain.Target)
+		target.Tag = ""
+	}
+
+	skipDB, err := bk.NewBuildkitSkipper(a.cli.Flags().LocalSkipDB, orgName, target, client)
 	if err != nil {
 		return nil, nil, false, err
 	}
 
+	targetConsole := a.cli.Console().WithPrefix(target.String())
+	targetStr := targetConsole.PrefixColor().Sprintf("%s", target.StringCanonical())
 	exists, err := skipDB.Exists(ctx, targetHash)
 	if err != nil {
-		console.Warnf("unable to check if target %s (hash %x) has already been run: %s", target.String(), targetHash, err.Error())
+		console.Warnf("Unable to check if target %s (hash %x) has already been run: %s", targetStr, targetHash, err.Error())
 		return nil, nil, false, nil
 	}
 
 	if exists {
-		consoleNoPrefix.Printf("target %s (hash %x) has already been run; exiting", target.String(), targetHash)
+		console.Printf("Target %s (hash %x) has already been run. Skipping.", targetStr, targetHash)
+		consoleNoPrefix.PrintSuccess()
 		return nil, nil, true, nil
 	}
 
@@ -926,7 +936,13 @@ func (a *Build) logShareLink(ctx context.Context, cloudClient *cloud.Client, tar
 	return logstreamURL, true, printLinkFn
 }
 
-func (a *Build) maybePrintBuildMinutesInfo(orgName string) {
+func (a *Build) maybePrintBuildMinutesInfo(cliCtx *cli.Context) {
+	orgName := a.cli.OrgName()
+	settings := a.cli.Flags().BuildkitdSettings
+	if !a.cli.IsUsingSatellite(cliCtx) || !settings.SatelliteIsManaged {
+		return
+	}
+
 	plan := billing.Plan()
 	if plan.GetMaxBuildMinutes() == 0 {
 		return