+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
 
 ## Unreleased
 
+## v0.7.23 - 2023-12-18
+
+### Added
+- Auto-skip (*experimental*) - a feature that allows you to skip large parts of a build in certain situations, especially suited for monorepos. For more information see [the auto-skip section from Caching in Earthfiles](https://docs.earthly.dev/docs/caching/caching-in-earthfiles#auto-skip).
+- A warning when a `COPY` destination includes a tilde (~). Related to [#1789](https://github.com/earthly/earthly/issues/1789).
+- A hint message to suggest the usage of `-i` flag to debug the build when a RUN command fails.
+- `start-interval` flag to `HEALTHCHECK` command for dockerfile parity [#3409](https://github.com/earthly/earthly/issues/3409).
+- A verbose message indicating which authentication providers are used during a build.
+- `ssh_command` config option which can be used to override the ssh command that is used by `git` when connecting to an ssh-based repository. Thanks to [@weaversam8](https://github.com/weaversam8) for the contribution!
+
+### Fixed
+- Limit the number of deprecation warnings when using `COMMAND` instead of `FUNCTION` keyword.
+- Fixed an error which stated `VERSION 0.0` is a valid Earthfile version.
+
+### Changed
+- Changed the color used to print metadata values (such as ARGs values) in the build log to Faint Blue.
+- Updated default alpine/git image to v2.40.1.
+- When creating an auth token, an existing token will no longer be overwritten by default. To overwrite, the `--overwrite` flag should be used.
+
+### Additional Info
+- This release includes changes to buildkit
+
 ## v0.7.22 - 2023-11-27
 
 ### Added